Added basic `p5.asciify` snippet for use without hooks.
From my perspective the most live-codeable way to use `p5.asciify` without relying on hard compiles, while allowing for even more customization, including the integration of other fancy p5.js add-on libraries and more.

Also acts as a simple example. This snippet is expected to be applied onto a 'new'  sketch with an empty `draw()` function and a setup function with just the `createCanvas()` call, with the canvas mode set to `WEBGL`. With this setup, the snippet can be applied and starts running immediately.